tried to add latext to pretext
